Rename Network effect to Communication, add monad transformers support
